chore: update profile layout
