+++ src/app/(main)/dashboard/profile/_components/profile-data.ts
@@ -73,12 +73,12 @@ export const profile: ProfileRecord = {
   personalEmail: "arhamkhnz@gmail.com",
   workPhone: "+1 (415) 555-0148",
   location: "San Francisco, CA",
-  workplace: "Hybrid · Mission office",
-  timezone: "Pacific Time (UTC−7)",
+  workplace: "Remote",
+  timezone: "UTC+5:30",
   employeeId: "ST-0284",
   startDate: "March 18, 2022",
   tenure: "4 years, 4 months",
-  employmentType: "Full-time employee",
+  employmentType: "Contractor",
   weeklyHours: "40 hours",
   schedule: "Monday–Friday · 9:00 AM–5:30 PM",
   probationEnd: "Completed June 18, 2022",
